VIRTUAL_ROOMBA: Unworking RoombaVirtual + relocated rest of consts
- Created the RoombaVirtual class. It currently simply replaces the
  RoombaRoowifi when created. Doesn't work yet.
- Moved the rest of constants into uiUtils.h and croiUtil.h

+++ src/croi/roombaRoowifi.h
@@ -26,8 +26,8 @@ public:
     virtual char getTemperature();
     virtual unsigned short getChargeLevel();
 
-    virtual short getDistance();
-    virtual short getAngle();
+    virtual double getDistance();
+    virtual double getAngle();
     virtual bool getLeftBumb();
     virtual bool getRightBumb();
